Mention Consent-O-Matic extension
The purpose of GDPR banners are not that you adblock it and never approve or review it. It is that you review and approve what you individually need per site and adblocking it will not solve it, it in fact can cripple the website and the futures it offers.

While lots of people just block, remove of hide the GDPR banners via adblocker this extension actually tries to automate the approval procedure.